150722 by jwiele on 2021/06/28 (pair: jshimkus) Merge 150333 from linux-uds to krusty. 150333 by jwiele on 2021/06/15 (pair: msakai) UDSUP-1 Remove PRI* print macros.
	short and unsigned short are guaranteed to be at least 16 bits
	wide. Therefore, int16_t can be cast to short and printed with
	"%h" and uint16_t can be cast to unsigned short and printed
	with "%hu".

	int and unsigned int are guaranteed to be at least 32 bits
	wide. Therefore, int32_t can be cast to int and printed with
	"%d".  uint32_t can be cast to unsigned int and printed with
	"%u".

	long and unsigned long are guaranteed to be at least 32 bits
	wide and can be printed with "%ld" and "%lu" respectively.

	long long int and unsigned long long int are guaranteed to be
	at least 64 bits wide. Therefore, int64_t can be cast to long
	long int and printed with "%lld" and uint64_t can be cast to
	unsigned long long int and printed with "%llu".

	The width of a pointer is implementation defined. It can be
	printed in the kernel with "%px" in production code or "%pK"
	in test code.

	In principal, uintmax_t is the C99 type of the largest
	unsigned integer type on the platform. It can be printed with
	"%j". Unfortunately, uintmax_t is undefined in the kernel
	build and printk does not support the "%j" format. However,
	the biggest integer type we use is uint64_t so we can achieve
	the same result by using unsigned long long int and printing
	it with "%llu" in the kernel module.

	size_t is implementation defined but guaranteed to be wide
	enough to represent the size in bytes of any object in memory
	and can be printed with "%zu". It is not guaranteed to be be
	big enough to represent the size of an object in bits. The
	size in bits of any object used by UDS can be printed safely
	with "%llu".

	.../typeDefs.h
	Removed definitions of PRI* print macros and typedef
	  uintmax_t.

	.../logger.h
	Replaced the macros for printing pointers with a single
	  preprocesor conditional in the only place they were used,
	  which was in remove_vmalloc_block.

	.../chapterIndex.c
	.../config.c
	.../deltaMemory.c
	.../index.c
	.../indexLayout.c
	.../indexPageMap.c
	.../indexZone.c
	.../masterIndex005.c
	.../volume.c
	Replaced print macros with explicit formats and argument types
	  as outlined above.

150734 by jshimkus on 2021/06/28 (pair: msakai)
Merge 150420 to krusty.
150420 by jshimkus on 2021/06/18 (pair: jwiele)
[VDOUP-6] Disambiguate UDS index layout functions.

	.../indexLayout.h
	.../indexLayout.c
	.../indexLayoutLinuxUser.c
	Disambiguated functions.

	.../index.c
	.../indexComponent.c
	.../indexInternals.c
	.../indexState.c
	.../udsMain.c
	.../volume.c
	.../volumeStore.c
	Adapted to UDS index layout disambiguation.

150737 by jshimkus on 2021/06/28 (pair: msakai)
Merge 150492 to krusty.
150492 by jshimkus on 2021/06/21 (pair: jwiele)
[VDOUP-6] Disambiguate UDS io factory functions.

	.../ioFactory.h
	.../ioFactoryLinuxUser.c
	Disambiguated functions.

	.../indexLayout.c
	.../indexLayoutLinuxUser.c
	.../fileIORegion.c
	.../bufferedReader.c
	.../bufferedWriter.c
	Adapted to io factory functions disambiguation.

150748 by jshimkus on 2021/06/29 (pair: msakai)
Merge 150545 to krusty.
150545 by jshimkus on 2021/06/22 (pair: msakai)
[UDSUP-1] Disambiguate UDS logger constants and functions.

	.../logger.h
	.../logger.c
	.../loggerLinuxUser.c
	Disambiguated constants and functions.

	.../fileIORegion.c
	.../fileUtils.c
	.../indexLayoutLinuxUser.c
	.../memoryLinuxUser.c
	.../minisyslog.c
	.../requestQueueUser.c
	.../syscalls.h
	.../threadsLinuxUser.c
	.../bufferedReader.c
	.../bufferedWriter.c
	.../chapterWriter.c
	.../deltaIndex.c
	.../deltaMemory.c
	.../hashUtils.c
	.../index.c
	.../indexCheckpoint.c
	.../indexComponent.c
	.../indexConfig.c
	.../indexInternals.c
	.../indexLayout.c
	.../indexLayoutParser.c
	.../indexPageMap.c
	.../indexRouter.c
	.../indexSession.c
	.../indexState.c
	.../indexStateData.c
	.../indexZone.c
	.../masterIndex005.c
	.../masterIndex006.c
	.../masterIndexOps.c
	.../openChapter.c
	.../openChapterZone.c
	.../pageCache.c
	.../permassert.c
	.../searchList.c
	.../stringUtils.c
	.../udsMain.c
	.../volume.c
	.../volumeStore.c
	Adapted to logger disambiguation.

150755 by jshimkus on 2021/06/29 (pair: msakai)
Adapt VDO to krusty change 150748.

	[UDSUP-1] Disambiguation of UDS constants, variables and functions.

	.../fixedLayout.c
	.../header.c
	.../superBlockCodec.c
	.../vdoComponentStates.c
	.../volumeGeometry.c
	.../fileLayer.c
	.../vdoConfig.c
	Adapted to krusty change 150748.

+++ utils/uds/chapterIndex.c
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
  * 02110-1301, USA. 
  *
- * $Id: //eng/uds-releases/krusty/src/uds/chapterIndex.c#13 $
+ * $Id: //eng/uds-releases/krusty/src/uds/chapterIndex.c#14 $
  */
 
 #include "chapterIndex.h"
@@ -136,8 +136,8 @@ int put_open_chapter_index_record(struct open_chapter_index *open_chapter_index,
 	found = was_entry_found(&entry, address);
 	result = ASSERT_WITH_ERROR_CODE(!(found && entry.is_collision),
 					UDS_BAD_STATE,
-					"Chunk appears more than once in chapter %" PRIu64,
-					open_chapter_index->virtual_chapter_number);
+					"Chunk appears more than once in chapter %llu",
+					(unsigned long long) open_chapter_index->virtual_chapter_number);
 	if (result != UDS_SUCCESS) {
 		return result;
 	}
@@ -193,8 +193,8 @@ int pack_open_chapter_index_page(struct open_chapter_index *open_chapter_index,
 		if (removals == 0) {
 			struct delta_index_stats stats;
 			get_delta_index_stats(delta_index, &stats);
-			uds_log_warning("The chapter index for chapter %" PRIu64 " contains %ld entries with %ld collisions",
-					open_chapter_index->virtual_chapter_number,
+			uds_log_warning("The chapter index for chapter %llu contains %ld entries with %ld collisions",
+					(unsigned long long) open_chapter_index->virtual_chapter_number,
 					stats.record_count,
 					stats.collision_count);
 		}
@@ -227,8 +227,8 @@ int pack_open_chapter_index_page(struct open_chapter_index *open_chapter_index,
 		} while (!entry.at_end);
 	}
 	if (removals > 0) {
-		uds_log_warning("To avoid chapter index page overflow in chapter %" PRIu64 ", %u entries were removed from the chapter index",
-				open_chapter_index->virtual_chapter_number,
+		uds_log_warning("To avoid chapter index page overflow in chapter %llu, %u entries were removed from the chapter index",
+				(unsigned long long) open_chapter_index->virtual_chapter_number,
 				removals);
 	}
 	return UDS_SUCCESS;
